requirements.txt
bcrypt==3.1.6
cffi==1.11.5
click==6.7
Flask==0.11.1
Flask-SQLAlchemy==2.3.2
jsonify==0.5
MarkupSafe==1.0
mysql-connector-python==8.0.11
pycparser==2.18
requests==2.22.0
